+++ QEfficient/transformers/models/gpt_bigcode/modeling_gpt_bigcode.py
@@ -439,7 +439,7 @@ class QEffGPTBigCodeForCausalLM(GPTBigCodeForCausalLM):
         # Cast to INT32 to avoid issue while running in ONNXRT
         logit_index = position_ids.to(torch.int32).argmax(1, keepdim=True)
         hidden_states = transformer_outputs[0][torch.arange(position_ids.shape[0]).view(-1, 1), logit_index]
-        lm_logits = self.lm_head(hidden_states)
+        lm_logits = self.lm_head(hidden_states).float()
 
         return CausalLMOutputWithCrossAttentions(
             loss=None,
